+++ sessoesdevoto/src/main/java/com/desafio/sessoesdevoto/dominio/dto/PautaSimplificadaDTO.java
@@ -2,6 +2,13 @@ package com.desafio.sessoesdevoto.dominio.dto;
 
 import com.desafio.sessoesdevoto.dominio.Pauta;
 
+/**
+ * Retorno com dados simplificados de uma pauta
+ *
+ * @param idPauta
+ * @param nome
+ * @param descricao
+ */
 public record PautaSimplificadaDTO(
         String idPauta,
         String nome,
