+++ sessoesdevoto/src/main/java/com/desafio/sessoesdevoto/dominio/dto/VotoCompletoDTO.java
@@ -2,6 +2,13 @@ package com.desafio.sessoesdevoto.dominio.dto;
 
 import com.desafio.sessoesdevoto.dominio.Voto;
 
+/**
+ * Retorno completo de um voto
+ *
+ * @param idDaPauta
+ * @param cpfAssociado
+ * @param voto
+ */
 public record VotoCompletoDTO(
         String idDaPauta,
         String cpfAssociado,
