+++ sessoesdevoto/src/main/java/com/desafio/sessoesdevoto/dominio/dto/VotoSimplificadoDTO.java
@@ -2,6 +2,12 @@ package com.desafio.sessoesdevoto.dominio.dto;
 
 import com.desafio.sessoesdevoto.dominio.Voto;
 
+/**
+ * Retorno simplificado de um voto
+ *
+ * @param cpfDoAssociado
+ * @param voto
+ */
 public record VotoSimplificadoDTO(
         String cpfDoAssociado,
         boolean voto
