+++ sessoesdevoto/src/main/java/com/desafio/sessoesdevoto/dominio/form/RegistrarPautaForm.java
@@ -1,4 +1,4 @@
-package com.desafio.sessoesdevoto.dominio.dto;
+package com.desafio.sessoesdevoto.dominio.form;
 
 import com.desafio.sessoesdevoto.dominio.Pauta;
