+++ sessoesdevoto/src/main/java/com/desafio/sessoesdevoto/dominio/portas/interfaces/VotoService.java
@@ -1,6 +1,6 @@
 package com.desafio.sessoesdevoto.dominio.portas.interfaces;
 
-import com.desafio.sessoesdevoto.dominio.dto.RegistrarVotoForm;
+import com.desafio.sessoesdevoto.dominio.form.RegistrarVotoForm;
 import com.desafio.sessoesdevoto.dominio.dto.VotoCompletoDTO;
 import com.desafio.sessoesdevoto.dominio.dto.VotoSimplificadoDTO;
